+++ src/datasets/segmentation/forward/shapenet.py
@@ -12,7 +12,7 @@ from src.core.data_transform import SaveOriginalPosId
 from src.utils import is_list
 from src.datasets.base_dataset import BaseDataset
 from src.metrics.shapenet_part_tracker import ShapenetPartTracker
-from .shapenet import ShapeNet
+from src.datasets.segmentation.shapenet import ShapeNet
 
 log = logging.getLogger(__name__)
